Refactor file structure and small add small patches for Windows
